make chain shared
