+++ examples/tokenvm/go.mod
@@ -3,11 +3,10 @@ module github.com/ava-labs/hypersdk/examples/tokenvm
 go 1.20
 
 require (
-	github.com/ava-labs/avalanche-network-runner v1.7.1-0.20230720220603-07c999585182
+	github.com/ava-labs/avalanche-network-runner v1.7.1
 	github.com/ava-labs/avalanchego v1.10.5
 	github.com/ava-labs/hypersdk v0.0.1
 	github.com/fatih/color v1.13.0
-	github.com/manifoldco/promptui v0.9.0
 	github.com/onsi/ginkgo/v2 v2.8.1
 	github.com/onsi/gomega v1.26.0
 	github.com/prometheus/client_golang v1.16.0
@@ -77,6 +76,7 @@ require (
 	github.com/kr/pretty v0.3.1 // indirect
 	github.com/kr/text v0.2.0 // indirect
 	github.com/magiconair/properties v1.8.6 // indirect
+	github.com/manifoldco/promptui v0.9.0 // indirect
 	github.com/mattn/go-colorable v0.1.13 // indirect
 	github.com/mattn/go-isatty v0.0.17 // indirect
 	github.com/mattn/go-runewidth v0.0.9 // indirect
